Two new lints

+++ server/src/game/player/player_accessors.rs
@@ -109,9 +109,9 @@ impl PlayerReference{
     pub fn player_tags<'a>(&self, game: &'a Game) -> &'a VecMap<PlayerReference, Vec1<Tag>>{
         &self.deref(game).player_tags
     }
-    pub fn player_has_tag(&self, game: &Game, key: PlayerReference, value: Tag) -> u8{
+    pub fn player_has_tag(&self, game: &Game, key: PlayerReference, value: Tag) -> usize {
         if let Some(player_tags) = self.deref(game).player_tags.get(&key){
-            player_tags.iter().filter(|t|**t==value).count() as u8
+            player_tags.iter().filter(|t|**t==value).count()
         }else{
             0
         }
